Add files via upload
This code simulates the function of a coffee distributor. It is based off of a UML class diagram. It has 4 buttons (0,1,2 and 3) that are used to customise your coffee, it takes in coins and has an admin where that lets you refill the sugar and the cups (gobelets). The admin password is 3012.